requirements.txt
transformers>=4.40.0
torch>=2.1.0
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.26.0
